backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ollama==0.4.4
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
psutil==6.1.1
pydantic==2.10.4
pydantic-settings==2.2.1
pydantic_core==2.27.2
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.6.0
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==12.0
